refactor: 매퍼 클래스 생성 및 로직 분리
