+++ src/main/java/com/speaktext/backend/book/voice/vo/CumulativeVoiceDuration.java
@@ -1,0 +1,14 @@
+package com.speaktext.backend.book.voice.vo;
+
+public class CumulativeVoiceDuration {
+
+    private final String cumulativeDurationsJson;
+
+    public CumulativeVoiceDuration(String cumulativeDurationsJson) {
+        this.cumulativeDurationsJson = cumulativeDurationsJson;
+    }
+
+    public String getJson() {
+        return cumulativeDurationsJson;
+    }
+}
